Refactor: use size_t and ptrdiff_t instead of int

+++ application.h
@@ -125,9 +125,9 @@ private:
 	sf::Vector2f world_dir_to_screenf(b2Vec2 world_dir);
 	b2Fixture* get_fixture_at(sf::Vector2i screen_pos);
 	GameObject* get_object_at(sf::Vector2i screen_pos);
-	int mouse_get_ground_vertex();
-	int mouse_get_ground_edge();
-	int mouse_get_edge_vertex();
+	ptrdiff_t mouse_get_ground_vertex();
+	ptrdiff_t mouse_get_ground_edge();
+	ptrdiff_t mouse_get_edge_vertex();
 	void select_vertices_in_rect(const RectangleSelect& rectangle_select);
 	void select_objects_in_rect(const RectangleSelect& rectangle_select);
 	void render_rectangle_select(sf::RenderTarget& target, RectangleSelect& rectangle_select);
